pyproject.toml
[project]
name = "g-voiceagent-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[google]~=1.0",
]
